requirements.txt
reflex==0.9.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
resend>=2.0.0